PR-SUG: move logging outside extensions
